@@ -2,7 +2,8 @@ package com.omegafrog.My.piano.app.user.vo;
 
 import lombok.Builder;
 
-import javax.persistence.Embeddable;
+import jakarta.persistence.Embeddable;
+import lombok.Getter;
 
 @Embeddable
 public class AlarmProperties {
